完善消息服务的设计

+++ model/src/main/java/com/tiktok/model/vo/user/FriendUser.java
@@ -12,5 +12,5 @@ public class FriendUser extends UserVo {
 
     String massage;
 
-    Long msgType;
+    Integer msgType;
 }
